Add Bentham et al's 2015 lupus study (#48)
- I have added the initial tasks for the Bentham et al.'s GWAS of Lupus.
- This includes tasks to pull data (both original build 37 data, and
data after liftover to build 38), convert it to GWASLAB sumstats form,
and produce region plots around the most significant variants.
- In the future, we can compare this study with the DecodeME data, via
genetic correlation and other approaches.
- I also added a ReadMe linking to the study itself.
- Update `GwasLabRegionPlotsFromLeadVariantsTask` to allow only plotting
the variants with smallest p values. IT is neccesary to use this option
for the Bentham et al GWAS, since it has so many significant variants.
In the future, we could consider parallelizing the creation of region
plots, if we think there is value in doing so.
- This PR closes #34