Add Pipe transformations, tasks to read RNA seq data (#95)
- Fuma's [SNP2GENE](https://fuma.ctglab.nl/tutorial#snp2gene) uses
RNASeq data to associate genes with tissues. I am considering
replicating and modifying this procedure locally. To do this, I need to
download and preprocess RNASeq data. This PR adds tasks to get RNASeq
data, as well as a new task type that uses DataProcessingPipes to
preprocess the RNASeq data.
- The initial preprocessing I am implementing is based on the
description on the FUMA website.